online mobile Billing
Mobile Billing system is used to manage the details of
connections types,Bills,connections,Tarrifs,customers. It
manages all the information about connections
types,payments,customers.
The purpose of this project is to built the website to reduce
the manual work for managing the connection
types ,Bills ,Payments ,connections. It tracks all the details
about the connections, Tariffs ,Customers.